Updated Readme.md - fixed maven-dependency
Changed dependecy-type from "pom" to "compile" - when using pom it will only add the dependecies declared in the meilisearch-java-pom, but not the SDK itself.
Below the definition for gradle uses "implementation", this means "compile" in Maven-Slang. To afford the same goal (use the SDK in the project), it should be compile, not only pom.